refactor: split GUI command palette
